# Improve performance of `object_has_rownames()` 