--- a/R/utils_object.R
+++ b/R/utils_object.R
@@ -84,5 +84,5 @@ object_has_rownames <- function(x) {
     stop("Only dataframe objects are allowed.", call. = FALSE)
   }
 
-  !identical(as.character(seq_len(nrow(x))), rownames(x))
+  !all(is.numeric(attributes(x)$row.names))
 }