@bwiernik 2022-09-30
Hmmm, I'm not sure this is strict enough though. For example, `psych::bfi` has numeric rownames that are participant IDs. How much performance cost is there to also check if the first element is `1` or the first few are `1:6`?