@etiennebacher 2022-09-30
Right, I'll make a new PR to take this into account